backend/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
aiosqlite==0.20.0
annotated-types==0.7.0
anthropic==0.45.0
anyio==4.8.0
attrs==25.1.0
backoff==2.2.1
beautifulsoup4==4.12.3
certifi==2024.12.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==41.0.7
dataclasses-json==0.6.7
defusedxml==0.7.1
deprecation==2.1.0
distro==1.9.0
docx2txt==0.8
emoji==2.14.1
et_xmlfile==2.0.0
eval_type_backport==0.2.2
fastapi==0.109.0
filetype==1.2.0
frozenlist==1.5.0
gotrue==2.11.2
greenlet==3.1.1
h11==0.14.0
h2==4.1.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
hyperframe==6.1.0
idna==3.10
jiter==0.8.2
joblib==1.4.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
langchain==0.3.15
langchain-anthropic==0.3.4
langchain-community==0.3.15
langchain-core==0.3.31
langchain-openai==0.3.2
langchain-text-splitters==0.3.5
langdetect==1.0.9
langgraph==0.2.67
langgraph-checkpoint==2.0.10
langgraph-checkpoint-sqlite==2.0.3
langgraph-sdk==0.1.51
langsmith==0.3.1
lxml==5.3.0
marshmallow==3.26.0
msgpack==1.1.0
multidict==6.1.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
nltk==3.9.1
numpy==2.2.2
openai==1.60.1
openpyxl==3.1.5
orjson==3.10.15
packaging==24.2
PyJWT==2.8.0
python-jose==3.3.0
passlib==1.7.4
postgrest==0.19.3
propcache==0.2.1
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic-core==2.27.2
pypdf==5.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-iso639==2025.1.27
python-magic==0.4.27
PyYAML==6.0.2
python-multipart==0.0.6
email-validator==2.1.0
bcrypt==4.1.2
RapidFuzz==3.11.0
realtime==2.2.0
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.37
starlette==0.45.3
storage3==0.11.1
StrEnum==0.4.15
supabase==2.3.0
supafunc==0.9.2
tabulate==0.9.0
tavily-python==0.5.0
tenacity==9.0.0
tiktoken==0.8.0
tqdm==4.67.1
typing-inspect==0.9.0
typing_extensions==4.12.2
unstructured==0.11.8
unstructured-client==0.29.0
urllib3==2.3.0
uvicorn==0.27.0
websockets==13.1
wrapt==1.17.2
yarl==1.18.3
zstandard==0.23.0
fredapi==0.5.2
